+++ src/main/java/com/cre/springbootinit/utils/LoginUserInfoUtil.java
@@ -3,9 +3,9 @@ package com.cre.springbootinit.utils;
 import java.util.Map;
 
 public class LoginUserInfoUtil {
-    public static Integer getUserId() {
+    public static Long getUserId() {
         // 获取当前登录用户的id
         Map<String, Object> map = ThreadLocalUtil.get();
-        return (Integer) map.get("id");
+        return (Long) map.get("id");
     }
 }
